Save the label by and color by options when bookmarking. Change: 134441940
